Factor out gridworld tests.
